ref: deduplicate isValidElement util function implementation
